requirements.txt
Django==5.2.7
Pillow==12.0.0
django-allauth==66.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
gunicorn==23.0.0

